refactor: move auth docs to account folder
